pyproject.toml
[project]
name = "back-test"
version = "0.1.0"
description = ""
authors = [
    {name = "LucasKim4321",email = "xowls0131@naver.com"}
]
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "certifi (>=2025.1.31,<2026.0.0)",
    "django-storages (>=1.14.6,<2.0.0)",
    "boto3 (>=1.37.27,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
django = "^5.1.7"
psycopg2-binary = "^2.9.10"
drf-spectacular = "^0.28.0"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = "^5.5.0"
django-extensions = "^3.2.3"
ipython = "^9.0.2"
certifi = "^2025.1.31"

